можно рисовать все фигуры кроме линии

+++ lab4/Lab4Task1Var2/Lab4Task1Var2/Circle.h
@@ -5,7 +5,7 @@
 class CCircle final : public ISolidShape
 {
 public:
-	CCircle(std::shared_ptr<CDot> &centerCircle, double &radius, std::string const & lineColor, std::string const & fillColor);
+	CCircle(std::shared_ptr<CDot> &centerCircle, float &radius, std::string const & lineColor, std::string const & fillColor);
 
 	double GetAreaShape() const override;
 	double GetPerimeterShape() const override;
